refactor: move Codex core logic into codex package
