+++ packages/codex/src/bridge.test.ts
@@ -3,7 +3,7 @@ import { EventEmitter } from "node:events";
 import { PassThrough } from "node:stream";
 import { describe, expect, it, vi } from "vitest";
 
-import { CodexBridge, type CodexMessage } from "./codex-bridge";
+import { CodexBridge, type CodexMessage } from "./bridge.ts";
 
 type WrittenMessage = {
   id?: number | string;
